Use mixin for translation things.
